Move BitcoindRpcAppConfig into the bitcoind-rpc project
