+++ app/server/src/main/scala/org/bitcoins/server/BitcoinSAppConfig.scala
@@ -12,6 +12,7 @@ import org.bitcoins.dlc.node.config.DLCNodeAppConfig
 import org.bitcoins.dlc.wallet.DLCAppConfig
 import org.bitcoins.keymanager.config.KeyManagerAppConfig
 import org.bitcoins.node.config.NodeAppConfig
+import org.bitcoins.rpc.config.BitcoindRpcAppConfig
 import org.bitcoins.tor.config.TorAppConfig
 import org.bitcoins.wallet.config.WalletAppConfig
