+++ node/src/main/scala/org/bitcoins/node/networking/peer/PeerMessageReceiver.scala
@@ -1,12 +1,13 @@
 package org.bitcoins.node.networking.peer
 
 import akka.actor.ActorRefFactory
+import org.bitcoins.core.api.node.NodeType
 import org.bitcoins.core.p2p._
 import org.bitcoins.node.config.NodeAppConfig
 import org.bitcoins.node.models.Peer
 import org.bitcoins.node.networking.P2PClient
 import org.bitcoins.node.networking.peer.PeerMessageReceiverState._
-import org.bitcoins.node.{Node, NodeType, P2PLogger}
+import org.bitcoins.node.{Node, P2PLogger}
 
 import scala.concurrent.Future
