+++ testkit/src/main/scala/org/bitcoins/testkit/node/NodeUnitTest.scala
@@ -5,6 +5,7 @@ import org.bitcoins.chain.blockchain.ChainHandlerCached
 import org.bitcoins.chain.config.ChainAppConfig
 import org.bitcoins.chain.models._
 import org.bitcoins.core.api.chain.ChainApi
+import org.bitcoins.core.api.node.NodeType
 import org.bitcoins.node._
 import org.bitcoins.node.config.NodeAppConfig
 import org.bitcoins.node.models.Peer
